pyproject.toml
[project]
name = "newsolingo"
version = "0.1.0"
description = "A language learning CLI that fetches real-world articles and creates CEFR-adapted reading exercises"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "trafilatura>=1.6.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
newsolingo = "newsolingo.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["newsolingo"]
